rfc5965 parser now accepts historic Received-Date and Version: 0.1
The version header from this old format is actually not at all
compatible with a strict reading of rfc5965 which allows only for
integer version fields; it is defined as `[1-9][0-9]*` which has
no room for decimal points.

Relax the field to accept an arbitrary string.

If Arrival-Date is not found, try pulling the equivalent information
from the historic Received-Date field.