Add Dockerfile for minimal image build
This uses alpine to build and the docker scratch image for installation,
resulting in an image of 150MB in size, compared to the 251MB size of
our current ubuntu based image.

I think we can shrink things down a bit more; for example, the kcli
utility pulls in the message crate for some of its types, but ends up
linking in openssl and other deps.  Splitting out the types into a
message-types crate might help to reduce the size of that executable.
Similarly, the other executables seem to have more stuff linked in
than is strictly needed.