rocky: use individual executable release build
`cargo build --release` will build all packages with the full
combination of enabled features, resulting in executables such as `kcli`
getting linked with rocksdb and openssl.

When we build packages we want to avoid that. In the CI we were running
`cargo nextest run --release` thinking that it would avoid recompiling
some deps and that we could reuse those "release" binaries in the final
package without recompiling. That isn't true, so let's just run the
tests in debug mode (should be slightly faster to build), then build
the individual executables in release mode, so that we get slimmer
executables.

We'll see how this works for rocky, then apply similar changes to
the rest of CI.