ci: docs: try using podman instead of docker(!)
It's a PITA to bootstrap the docker daemon, and pointing it to the host
daemon to sidestep all that is insecure and should not be trusted in a
public repo like this.

So, let's just try podman for this: it doesn't need a daemon and
doesn't need any special privs.