tsa: add separate suspension concept
This is the start of unbundling and unwinding the mild hack
that was adding a suspended configuration state.  At the time
that was implemented, there was no plan in place for handling
stuff other than configuration updates for TSA rules.

I have a simple but powerful plan in mind, so let's start
breaking this back out!

refs: https://github.com/KumoCorp/kumomta/issues/113