add overall_max_message_rate option to queue helper
The intent is to limit the overall rate at which a tenant
is able to send messages out from the server. It is NOT related
to the rate at which messages can be injected to ths server.

This is allowed in two locations:

* Directly in the tenant block
* In a new campaigns table located under a tenant

```toml
[tenant.'mytenant']
egress_pool = 'tpool'
overall_max_message_rate = "100/s"

[tenant.'mytenant'.campaigns.'mycampaign']
overall_max_message_rate = "50/s"
```

The tenant level throttle is checked first, and if it permits
the message, then the campaign specific throttle is checked.

refs: https://github.com/KumoCorp/kumomta/issues/143