smtp client: don't unconditionally log bind errors
The thesis here was that you'd always want to see these errors in
the journal, but it is common for sites that have both A and AAAA
records report bind failures if the local machine has no IPv6
connectivity.  It's ok: we'll end up using an IPv4 address later
in the connection plan, and if there was ultimately a failure to
connect, that error will show up in the delivery logs.

In the common case, logging the error from this location in
the code is a false positive and is a noisy choice, so let's
not do that.