bug: math in ideal_connection_count yields too-big values
This isn't a proper fix, but a bandaid.  When the max connection limit
is large (eg: 1024), the calculation here can produce numbers larger
than the current queue size, which can result in surging to spawn
dispatchers which then realize that they have no work to do, on each
queue insertion operation.

This commit clamps the ideal number to be no larger than the current
queue size.

Will need to follow up on this with the correct math.

On more reasonable queue sizes (such as those covered by the unit
tests), the numbers are reasonable.