amortize ready-q maintainer triggers
Previously, every insertion into the ready queue would cause the
maintainer logic to run in the context of the submitting thread.

For busy, high-concurrency systems this could result in multiple
dispatches of the maintainer for the same queue in quick succession.

This commit adjusts the triggering logic so that the existing
background maintainer task will be woken up to perform those
actions.

This reduces the number of redundant calls to the maintainer,
without harming latency.