tsa: remove local re-resolve of site-name
The rationale was that this would be more robust, but as it turns out,
it may result in discrepenancies with the incoming site_name that
affect how we match back for suspensions.

I saw localhost getting over-resolved in the integration tests so
that it would no longer match.

Let's just KISS and trust the incoming data.