mailparsing: introduce PartPointer
This is vaguely reminiscent of a json pointer reference that
symbolically identifies a node in a tree using a sequence of
integers to indicate the child node index by tree level.

The simplified structure method has been re-implemented in terms
of PartPointer.

It is possible to resolve a PartPointer to its associate MimePart
for both mutable and immutable access.

refs: https://github.com/KumoCorp/kumomta/issues/117
refs: https://github.com/KumoCorp/kumomta/issues/120