fixup and fully migrate to metrics 0.22
Commit 12fe5e3b61552c94f54c129928bb8698890a49fc updated the
metrics related crates, but not in every one of our crates,
which meant that we were running with a mixture of 0.20
and 0.22.

The result of this was that the memory related metrics
were no long visible to the published prometheus data
because it was only looking at the other version of
the metrics crate!

This commit switches over to using a workspace dep
for metrics so that we update all related crates
together.