tsa: make the default queue config for tsa-daemon more aggressive
while troubleshooting a situation where the
http://127.0.0.1:8008.tsa.kumomta queue is backing up, I noticed that
we're using the default queue configuration for this queue.

Let's make it more inline with the defaults for webhooks; give it a
1 minute base retry with a 20 minute max.

These parameters are configurable; you can pass in `tsa_queue_config` to
the setup_with_automation call to specify your preferred values for the
scheduled queue config.