egress_source: add preliminary support for socks5 auth
This commit adds two new fields to the egress source definition:

* socks5_proxy_username - string
* socks5_proxy_password - keysource
  (https://docs.kumomta.com/reference/keysource/)

when present, the socks5 client will indicate that it is capable
of doing username/password authentication per
https://www.rfc-editor.org/rfc/rfc1928

The password is specified as a keysource, which allows it to
load credentials from a vault or a file in addition to a
string.

This is untested beyond compilation, so there are currently
no docs for this.