queue: lock scoping
It wasn't the first time and it won't be the last time that this
class of issue crops up :-/

This pattern:

```
match something.lock().pop() {
   ...
}
```

will hold the lock even after the first line has completed
and copied its data out.

This causes subtle and difficult to find issues where
the lock lifetime overlaps with some other operation.

The resoluion for this is to encapsulate the lock
operation inside a function to break the lifetime.