spool: optimize removal code path
Previously, we would spawn a new task to asynchronously,
wrt. to the delivery attempt, remove a message from storage.

What this meant was that we'd end up cloning a reference
to a Message, inflating the message count, as well as allocating
the resources needed to run that removal routine.

Since this was O(attempts) it resulted in putting a lot of
pressure on the tokio task queues.

Since the recommended rocksdb spool doesn't actually do any
material work for the remove operation (it is a short critical
section that doesn't block), all that async work was purely
overhead.

This commit removes it in favor of simply calling the remove
routine in the context where it is needed.

This results in a very stable and low memory utilization
without impacting performance.

While in there, shave off a couple of await operations
that aren't strictly needed around fetching spool handles.

Err, and in "this should really have been a separate commit",
centralize the notify-maintainer-on-Dispatcher completion
to the Drop handler. Previously we would only do this when
closing out the connection due to reaching the message limit.
We now do this always to give the maintainer a chance to spawn
additional connections whenever any of them close. Let's make
the wakeups for maintainers and dispatchers always use notify_one,
which reduces overall CPU utilization.