rocksdb: improve write/delete throughput
This adjusts how we do write and delete operations so that we
first try to immediately submit the operation to the WAL,
but if that would block, then we push the operation to
the tokio blocking thread pool in order not to hard-block
the tokio schedule thread when rocksdb needs to perform writes.

While in there, hook up the force-sync flag to the closest
approximation in the WriteOptions for the batch (of size 1).