suspensions: make looking up a readyq suspension O(1)
Previously, we'd be O(number-of-suspensions) and do some extra
book keeping work as part of every operation.

This commit reorganizes things so that the looking is now O(1)
via a hash table for the queue name (the original code was
a copy-pasta adaptation from the scheduled queue code that
could match based on multiple critiera, and couldn't use a hash lookup).

The expiration logic is made "probabalistic" (really, once every N
operations), and the caller can now hold a "reference" to the
suspension result to avoid locking to check to see if it is
still current.

I'm mostly happy with this, except that for egress source selection
it is not easy to cache without introducing an additional point
of contention to coordinate on maintaining the cache.