ci: ubuntu arm
For the docker image, ideally we'd build a multiarch image but
I'm unsure how best to persuade woodpecker-ci (our build infra)
to wait for both native architecture builds to complete before
scheduling the `docker manifest` build step to merge them
together.

For now we define `latest-arm64` as the way to select an arm
build, or `2024.06.04-5e530b2d-arm64` to select a specific
stable release version for arm.