lua_deliver: add close method
This is to facilitate more precise control over the lifetime of
connection objects.

You can now define `function sender:close()` on the connection
object that is returned from your lua protocol constructor.

This method will be called when the underlying
QueueDispatcher::close_connection method is called.

Adjust the logic in ready_queue to explicitly call
QueueDispatcher::close_connection in the cases where we're closing the
connection; previously we'd just leave it to the Drop handler to take
care of this.

Adjust the amqp docs to show how to bubble the close call through
to the associated client.