docs: make since() macro a bit smarter
Each time we add a new page in `main`, we get an INFO line like this,
because we didn't know the relative link:

INFO    -  Doc file 'reference/string/psl_domain.md' contains an absolute link '/userguide/installation/linux/', it was left as is. Did you mean
           '../../userguide/installation/linux.md'?

this commit adds some logic to compute the appropriate relative
path and make the doc build a bit less noisy.