add tls-probe utility
This is conceptually similar in purpose to running `openssl s_client`
to figure out the capabilities of a remote mx and its TLS
implementation.

This commit also adds a currently-deliberately-undocumented
tls_prefer_openssl option to the egress path config that forces
the selection of an openssl based client connector.

I want to try this out a bit before fleshing it out into a
larger product feature and its additional configuration.

RUST_LOG=trace ./tls-probe example.com:25