fix existential issue with queue helper and egress_pool in --validate mode
When running in validation mode, if a given queue block didn't specify
an egress_pool, we could hit a validation error when the validator
tried to access that missing field.

The issue is that while we can dynamically validate properties against
the underlying rust code on assignment (via newindex), we don't have
a way to validate a read (index) operation because the validator works
by creating a dummy value and serializing it.

We could probably fix this up by introducing a read validator that
tries to create an empty object and reading back its fields, but
we don't currently generate index/field getters for those rust
structs.

So for now we will allow any read of a field that isn't set
in a typed record table when it has a dynamic field validator
present.