rfc5321: borrow client tls setup from rustls mio example client
In theory, this should be a NOP change. Adjust the setup logic
for the rustls client config to more closely match the tls client
in
<https://github.com/rustls/rustls/blob/rel-0.21/examples/src/bin/tlsclient-mio.rs>

This makes it a bit easier to introduce configuration for the cipher
suite and other parameters in the future, as well as to upgrade to
a newer version of rustls in the future.