rfc5321 client: try harder to continue in clear after failed openssl handshake
The API for this in tokio-openssl and the underlying openssl crate are
both very awkward for this particular case.

The approach taken here is to dup the underlying socket (if it is
a stream that is a TcpStream) and keep that around to use for the
fallback case.

This isn't perfect; there are some failure modes where the peer
still has encoded binary data either buffered or where it thinks
the session is still setup.